pyproject.toml
[tool.poetry]
name = "erp"
version = "0.1.0"
description = ""
authors = ["Nicolás Guglielmi Manent <nicolas.guglielmi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
django-model-utils = "^4.3.1"
django-simple-history = "^3.4.0"
pillow = "^10.0.1"
django-jet-reboot = "^1.3.3"
six = "^1.16.0"
django-clone = "^5.3.3"
xhtml2pdf = "^0.2.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
